Graph2 (#3)
adds releases for the otl wheel -- currently we only compile for x86_64 linux, and arm64/x86 macos. builds are triggered by releases with the format x.y.z where x,y,and z must be base10 integers.
restructured to a cargo workspace -- if anything feels "orthogonal" or "shared", we should split it out to a new crate in the workspace. better for seperation of concerns, and also decreases compile times, as a crate is a single unit of compile
added a tap-out fn that can be used by the job runner to get scripts for each test
this also fixed a bunch of build fails with the wheel, had some issues where the python files werent being packaged

---------

Co-authored-by: James Connolly <james@silogy.io>